openos installer part 1 done

+++ installer/idea.lua
@@ -1,9 +1,16 @@
 --[[
     Boot Installer:
-        format drive
-        get all os files
-        copy os files (Rename init to osinit)
-        remove installer files
-        rename osinit to init
-        reboot
+        After Bottom:
+            ask if they really wanna continue and format
+            if they dont, say that it will now revert changes and if you want to just format then reboot
+            after that if they dont, revert changes
+
+            format drive
+
+        Start off with:
+            get all os files
+            copy os files (Rename init to osinit)
+            remove installer files
+            rename osinit to init
+            reboot
 ]]
